+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 * Changed solidify modifier to use the "OutlineThickness" vertex group for thickness.
 * Changed wimdo export to use the "OutlineThickness" vertex group for "OutlineVertexColor" alpha.
+* Adjusted wimdo export to report warnings if outlines cannot be generated.
 
 ### Fixed
 * Fixed an issue where some camdo models would not correctly remap material texture assignments.